--- a/benchmarks/src/main/scala/benchmark/Benchmark.scala
+++ b/benchmarks/src/main/scala/benchmark/Benchmark.scala
@@ -16,11 +16,11 @@
 
 package benchmark
 
-import scala.collection.mutable
 import java.net.URI
 import java.nio.file.{Files, Paths}
 import java.nio.charset.StandardCharsets
 
+import scala.collection.mutable
 import scala.language.postfixOps
 import scala.sys.process._
 import scala.util.control.NonFatal
@@ -33,6 +33,7 @@ import com.fasterxml.jackson.module.scala.{DefaultScalaModule, ScalaObjectMapper
 import org.apache.spark.SparkUtils
 import org.apache.spark.sql.{Row, SparkSession}
 import org.apache.spark.sql.functions.col
+import org.apache.spark.sql.internal.SQLConf
 
 trait BenchmarkConf extends Product {
   /** Cloud path where benchmark data is going to be written. */
@@ -111,6 +112,11 @@ abstract class Benchmark(private val conf: BenchmarkConf) {
     s
   }
 
+  val extraConfs: Map[String, String] = Map(
+    SQLConf.BROADCAST_TIMEOUT.key -> "7200",
+    SQLConf.CROSS_JOINS_ENABLED.key -> "true"
+  )
+
   private val queryResults = new mutable.ArrayBuffer[QueryResult]
   private val extraMetrics = new mutable.HashMap[String, Double]
 
@@ -157,7 +163,7 @@ abstract class Benchmark(private val conf: BenchmarkConf) {
         log(s"ERROR: $banner\n${e.getMessage}")
         queryResults +=
           QueryResult(queryName, iteration, durationMs = None, errorMsg = Some(e.getMessage))
-        if (!ignoreError) throw e else spark.empty
+        if (!ignoreError) throw e else Nil
     }
   }